Work

+++ backend/api/views.py
@@ -3,6 +3,7 @@ from collections import defaultdict
 from django.http import HttpResponse
 from django.shortcuts import get_object_or_404
 from rest_framework.response import Response
+from rest_framework.permissions import IsAuthenticated, AllowAny
 from rest_framework import viewsets, status, filters, permissions
 from django_filters.rest_framework import DjangoFilterBackend
 from rest_framework.decorators import action
@@ -10,7 +11,7 @@ from django.urls import reverse
 from rest_framework.pagination import LimitOffsetPagination
 from .filters import RecipeFilter
 
-from core.utils import User
+from core import utils
 from .serializers import (
     UserSerializer,
     UserRegistrationSerializer,
@@ -38,12 +39,14 @@ from .permissions import AuthorOrReadOnly
 
 
 class UserViewSet(viewsets.ModelViewSet):
-    queryset = User.objects.all()
+    """Вьюсет для обработки запросов с пользователями."""
+    queryset = utils.User.objects.all()
     serializer_class = UserSerializer
-    permission_classes = [permissions.AllowAny]
+    permission_classes = (AllowAny,)
     pagination_class = LimitOffsetPagination
 
     def create(self, request, *args, **kwargs):
+        """Создание польователя."""
         serializer = UserRegistrationSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
@@ -51,49 +54,60 @@ class UserViewSet(viewsets.ModelViewSet):
 
     @action(
         detail=False,
-        methods=['get'],
-        permission_classes=[permissions.IsAuthenticated]
+        methods=utils.get_method(),
+        permission_classes=[IsAuthenticated]
     )
     def me(self, request):
-        user = request.user
+        """
+        Получение пользователем своего профиля. Доступно только
+        аутентифицированным пользователям.
+        """
+        user = utils.get_user(request)
         serializer = self.get_serializer(instance=user)
         return Response(serializer.data)
 
     @action(
         detail=False,
-        methods=['put', 'delete'],
+        methods=utils.put_delete_methods(),
         url_path='me/avatar',
-        permission_classes=[permissions.IsAuthenticated]
+        serializer_class=AvatarSerializer,
+        permission_classes=[IsAuthenticated]
     )
     def avatar(self, request):
-        user = request.user
+        """Добавление/удаление аватара (фото профиля)."""
+        user = utils.get_user(request)
+
         if request.method == 'PUT':
-            user = request.user
-            serializer = AvatarSerializer(
-                user,
-                data=request.data, partial=True
+            serializer = self.get_serializer(
+                user, data=request.data, partial=True
             )
             serializer.is_valid(raise_exception=True)
             serializer.save()
             return Response(serializer.data, status=status.HTTP_200_OK)
 
-        avatar = user.avatar
+        if request.method == 'DELETE':
+            avatar = user.avatar
 
-        if avatar:
-            avatar.delete()
-            user.save()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+            if avatar:
+                avatar.delete()
+                user.save()
+
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
     @action(
         detail=False,
-        methods=['post'],
+        methods=utils.post_method(),
         serializer_class=ChangePasswordSerializer,
-        permission_classes=[permissions.IsAuthenticated]
+        permission_classes=[IsAuthenticated]
     )
     def set_password(self, request):
+        """
+        Смена пароля для авторизированного пользователя. Необходимо указать
+        текущий пароль и новый.
+        """
         serializer = self.get_serializer(data=request.data)
         serializer.is_valid(raise_exception=True)
-        user = request.user
+        user = utils.get_user(request)
         new_password = serializer.validated_data['new_password']
         user.set_password(new_password)
         user.save(update_fields=['password'])
@@ -101,47 +115,59 @@ class UserViewSet(viewsets.ModelViewSet):
 
     @action(
         detail=True,
-        methods=['post', 'delete'],
-        permission_classes=[permissions.IsAuthenticated]
+        methods=utils.post_delete_methods(),
+        permission_classes=[IsAuthenticated]
     )
     def subscribe(self, request, pk=None):
-        user = request.user
-        subscribe_user = get_object_or_404(User, pk=pk)
+        """Добавление/удаление подписки на пользователя."""
+        user = utils.get_user(request)
+        subscribe_user = get_object_or_404(utils.User, pk=pk)
 
         if request.method == 'POST':
             recipes_limit = request.query_params.get('recipes_limit')
-            serializer = SubscribeSerializer(
-                data={
-                    'user': user,
-                    'subscribing': subscribe_user.id
-                },
-                context={'request': request, 'recipes_limit': recipes_limit}
-            )
-            if serializer.is_valid():
-                serializer.save()
-                return Response(serializer.data, status=status.HTTP_201_CREATED)
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            data = {
+                'user': user, 'subscribing': subscribe_user.id
+            }
+            context = {'request': request, 'recipes_limit': recipes_limit}
+            serializer = SubscribeSerializer(data=data, context=context)
 
-        elif request.method == 'DELETE':
-            serializer = SubscribeDeleteSerializer(data={'subscribing_id': subscribe_user.id}, context={'request': request})
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                return Response(
+                    serializer.data, status=status.HTTP_201_CREATED
+                )
+
+            return Response(
+                serializer.errors, status=status.HTTP_400_BAD_REQUEST
+            )
+
+        if request.method == 'DELETE':
+            data = {'subscribing_id': subscribe_user.id}
+            context = utils.get_request(request)
+            serializer = SubscribeDeleteSerializer(data=data, context=context)
+
             if serializer.is_valid():
-                subscribe = get_object_or_404(Subscribe, user=user, subscribing=subscribe_user)
+                subscribe = get_object_or_404(
+                    Subscribe, user=user, subscribing=subscribe_user
+                )
                 subscribe.delete()
                 return Response(status=status.HTTP_204_NO_CONTENT)
-            
+
             errors = serializer.errors
 
             if 'not_found' in str(errors):
                 return Response(errors, status=status.HTTP_404_NOT_FOUND)
+
             return Response(errors, status=status.HTTP_400_BAD_REQUEST)
 
     @action(
         detail=False,
-        methods=['get'],
+        methods=utils.get_method(),
         permission_classes=[permissions.IsAuthenticated]
     )
     def subscriptions(self, request):
-        user = request.user
+        """Получение информации о своих подписках."""
+        user = utils.get_user(request)
         subscriptions = Subscribe.objects.filter(user=user).select_related(
             'subscribing'
         ).prefetch_related('subscribing__recipes')
@@ -154,71 +180,92 @@ class UserViewSet(viewsets.ModelViewSet):
 
 
 class RecipeList(viewsets.ModelViewSet):
+    """Вьюсет для обработки запросов с рецептами."""
     queryset = Recipe.objects.all()
     serializer_class = RecipeSerializer
-    permission_classes = [AuthorOrReadOnly]
+    permission_classes = (AuthorOrReadOnly,)
     pagination_class = LimitOffsetPagination
     filter_backends = (filters.OrderingFilter, DjangoFilterBackend)
     filterset_class = RecipeFilter
-    filterset_fields = ('author', 'tags')
+    filterset_fields = ('author',)
     ordering = ('-pub_date',)
-    http_method_names = ['get', 'post', 'patch', 'delete']
+    http_method_names = ('get', 'post', 'patch', 'delete')
 
     def get_queryset(self):
+        """
+        Получение списка рецептов в завизимости от наличия в списке покупок или
+        в избранном.
+        """
         queryset = super().get_queryset()
-        user = self.request.user
+        user = utils.get_self_user(self)
         is_in_card = self.request.query_params.get('is_in_shopping_cart')
         is_in_favorite = self.request.query_params.get('is_favorited')
 
         if not user.is_authenticated:
             if is_in_card == '1' or is_in_favorite == '1':
-                return queryset.none()
-            else:
                 return queryset
 
         if is_in_card is None and is_in_favorite is None:
             return queryset
         
         if is_in_card == str(1):
-            cart_recipe = ShoppingCard.objects.filter(user=user).values_list('recipe_id', flat=True)
+            cart_recipe = ShoppingCard.objects.filter(
+                user=user
+            ).values_list('recipe_id', flat=True)
             queryset = queryset.filter(id__in=cart_recipe)
-        if is_in_favorite == str(1):
-            favorite_recipe = ShoppingCard.objects.filter(user=user).values_list('recipe_id', flat=True)
+
+        elif is_in_favorite == str(1):
+            favorite_recipe = ShoppingCard.objects.filter(
+                user=user
+            ).values_list('recipe_id', flat=True)
             queryset = queryset.filter(id__in=favorite_recipe)
+
         return queryset
-        
-    
 
     def get_permissions(self):
+        """Доступ на редактирование только автору рецепта."""
         if self.action in ['list', 'retrieve']:
-            return [permissions.AllowAny()]
-        return [AuthorOrReadOnly()]
+            return (AllowAny(),)
+        return (AuthorOrReadOnly(),)
 
     def perform_create(self, serializer):
-        serializer.save(author=self.request.user)
+        """Присваивание автора рецепта текущего пользователя"""
+        serializer.save(author=utils.get_self_user(self))
     
     @action(
         detail=True,
-        methods=['post', 'delete'],
-        permission_classes=[permissions.IsAuthenticated]
+        methods=utils.post_delete_methods(),
+        permission_classes=[IsAuthenticated]
     )
     def favorite(self, request, pk=None):
-        user = request.user
+        """Добавление/удаление рецептов в избранное."""
+        user = utils.get_user(request)
+        data = utils.get_recipe_id(pk)
+        context = utils.get_request(request)
 
         if request.method == 'POST':
             recipe = get_object_or_404(Recipe, pk=pk)
-            serializer = FavoriteSerializer(data={'recipe': recipe.id}, context={'request': request})
-            if serializer.is_valid():
+            serializer = FavoriteSerializer(data=data, context=context)
+
+            if serializer.is_valid(raise_exception=True):
                 serializer.save()
-                return Response(serializer.data, status=status.HTTP_201_CREATED)
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+                return Response(
+                    serializer.data, status=status.HTTP_201_CREATED
+                )
+
+            return Response(
+                serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
         if request.method == 'DELETE':
-            serializer = FavoriteDeleteSerializer(data={'recipe_id': pk}, context={'request': request})
+            serializer = FavoriteDeleteSerializer(data=data, context=context)
+
             if serializer.is_valid():
                 recipe = serializer.context['recipe']
-                favorite = get_object_or_404(Favorite, user=user, recipe=recipe)
+                favorite = get_object_or_404(
+                    Favorite, user=user, recipe=recipe
+                )
                 favorite.delete()
+
                 return Response(status=status.HTTP_204_NO_CONTENT)
 
             errors = serializer.errors
@@ -230,36 +277,43 @@ class RecipeList(viewsets.ModelViewSet):
 
     @action(
         detail=True,
-        methods=['post', 'delete'],
-        permission_classes=[permissions.IsAuthenticated]
+        methods=utils.post_delete_methods(),
+        permission_classes=[IsAuthenticated]
     )
     def shopping_cart(self, request, pk=None):
-        user = request.user
+        """Добавление/удаление рецептов в список покупок."""
+        user = utils.get_user(request)
+        data = utils.get_recipe_id(pk)
+        context = utils.get_request(request)
 
         if request.method == 'POST':
             recipe = get_object_or_404(Recipe, pk=pk)
-            serializer = ShoppingCardSerializer(
-                data={'recipe': recipe.id},
-                context={'request': request}
-            )
-            if serializer.is_valid():
+            serializer = ShoppingCardSerializer(data=data, context=context)
+
+            if serializer.is_valid(raise_exception=True):
                 serializer.save()
-                return Response(serializer.data, status=status.HTTP_201_CREATED)
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+                return Response(
+                    serializer.data, status=status.HTTP_201_CREATED
+                )
+
+            return Response(
+                serializer.errors, status=status.HTTP_400_BAD_REQUEST
+            )
 
         if request.method == 'DELETE':
-            serializer = ShoppingCardDeleteSerializer(
-                data={'recipe_id': pk},
-                context={'request': request}
-            )
+            serializer = ShoppingCardDeleteSerializer(data=data, context=context)
 
             if serializer.is_valid():
                 recipe = serializer.context['recipe']
-                card = get_object_or_404(ShoppingCard, user=user, recipe=recipe)
-                card.delete()
+                favorite = get_object_or_404(
+                    ShoppingCard, user=user, recipe=recipe
+                )
+                favorite.delete()
+
                 return Response(status=status.HTTP_204_NO_CONTENT)
 
             errors = serializer.errors
+
             if 'not_found' in str(errors):
                 return Response(errors, status=status.HTTP_404_NOT_FOUND)
 
@@ -267,21 +321,26 @@ class RecipeList(viewsets.ModelViewSet):
 
     @action(detail=True, url_path='get-link')
     def get_link(self, request, pk=None):
+        """Получение короткой ссылки на рецепт."""
         recipe = get_object_or_404(Recipe, pk=pk)
         short_code = str(recipe.pk)
         relative_path = f'/api/recipes/{short_code}/'
         short_url = request.build_absolute_uri(relative_path)
         return Response({'short-link': short_url})
-    
+
     @action(
         detail=False,
-        methods=['get'],
-        permission_classes=[permissions.IsAuthenticated]
+        methods=utils.get_method(),
+        permission_classes=[IsAuthenticated]
     )
     def download_shopping_cart(self, request):
-        user = request.user
+        """Скачивание ингредиентов из списка покупок."""
+        user = utils.get_user(request)
 
-        shopping_card_for_user = ShoppingCard.objects.filter(user=user).select_related('recipe')
+        shopping_card_for_user = ShoppingCard.objects.filter(
+            user=user
+        ).select_related('recipe')
+
         recipes = [item.recipe for item in shopping_card_for_user]
 
         if not recipes:
@@ -302,16 +361,25 @@ class RecipeList(viewsets.ModelViewSet):
         for (name, unit), total in sorted_items:
             text_to_print += f'{name}: {total} {unit}\n'
         
-        data = [{'name': name, 'total': total, 'unit': unit} for (name, unit), total in ingredients.items()]
+        data = [
+            {'name': name, 'total': total, 'unit': unit}
+            for (name, unit), total in ingredients.items()
+        ]
 
         file_data = self.ingredients_to_text(data)
 
-        response = HttpResponse(file_data, content_type='text/plain; charset=utf-8')
-        response['Content-Disposition'] = 'attachment; filename="shopping_list.txt"'
+        response = HttpResponse(
+            file_data, content_type='text/plain; charset=utf-8'
+        )
+        response[
+            'Content-Disposition'
+        ] = 'attachment; filename="shopping_list.txt"'
+
         return response
 
     @staticmethod
     def ingredients_to_text(ingredients):
+        """Получение текстового файла с ингредиентами."""
         text_to_print = 'Нужно купить:\n'
         text_to_print += '\n'.join(
             f'{item['name']}: {item['total']} {item['unit']} '
@@ -321,6 +389,7 @@ class RecipeList(viewsets.ModelViewSet):
         return BytesIO(text_to_print.encode('utf-8'))
 
 class TagReadOnlyModelViewSet(viewsets.ReadOnlyModelViewSet):
+    """Вьюсет для обработки запросов с тегами."""
     queryset = Tag.objects.all()
     serializer_class = TagSerializer
     pagination_class = None
@@ -328,6 +397,7 @@ class TagReadOnlyModelViewSet(viewsets.ReadOnlyModelViewSet):
 
 
 class IngredientReadOnlyModelViewSet(viewsets.ReadOnlyModelViewSet):
+    """Вьюсет для обработки запросов с ингредиентами."""
     queryset = Ingredient.objects.all()
     serializer_class = IngredientSerializer
     pagination_class = None
